Working basic demo
A lot of this was supposed to be last commit.
Again too many changes, but it's working! I have a basic
parser, model, and editor/code generator working. You can change
either the canvas by dragging the circle, or change the text.
Both update the model and propogate back to the other
representation. It was way easier than I thought to get to this
point. There's still a lot of work to be done fleshing out the
language and adding functionality to the parser. And of course
actually making a useful canvas editor.