Flattened triangle vertices representation
Rather than being a nested attribute, the 3 vertices are now
defined by x1 y1 x2 y2 x3 and y3. This allows them to be parsed
the same as other attributes and I think looks cleaner in the
representation. It also follows the same pattern used by SVG
lines, which I'm planning to use for ANML lines.